test: GeoJSONスキーマテストを責務別に配置
