+++ nusamai/src/source/geojson_schema.rs
@@ -78,3 +78,126 @@ impl GeoJsonSchemaBuilder {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use serde_json::json;
+
+    fn properties(value: serde_json::Value) -> JsonObject {
+        value
+            .as_object()
+            .expect("test value must be an object")
+            .clone()
+    }
+
+    #[test]
+    fn collects_union_in_deterministic_order_with_optional_attributes() {
+        let first = properties(json!({"zeta": 1, "alpha": "first"}));
+        let second = properties(json!({"middle": true, "alpha": "second"}));
+        let mut builder = GeoJsonSchemaBuilder::default();
+
+        builder.observe_properties(Some(&first));
+        builder.observe_properties(None);
+        builder.observe_properties(Some(&second));
+        let feature_type = builder.finish();
+
+        assert_eq!(
+            feature_type
+                .attributes
+                .keys()
+                .map(String::as_str)
+                .collect::<Vec<_>>(),
+            ["alpha", "middle", "zeta"]
+        );
+        assert!(!feature_type.additional_attributes);
+        assert!(feature_type.attributes.values().all(|attribute| {
+            attribute.min_occurs == 0
+                && attribute.max_occurs == Some(1)
+                && attribute.original_name.is_none()
+        }));
+    }
+
+    #[test]
+    fn maps_each_json_value_kind_to_schema_type() {
+        let values = properties(json!({
+            "array": [1, 2],
+            "boolean": true,
+            "double": 1.5,
+            "integer": -1,
+            "object": {"nested": "value"},
+            "string": "001"
+        }));
+        let mut builder = GeoJsonSchemaBuilder::default();
+
+        builder.observe_properties(Some(&values));
+        let feature_type = builder.finish();
+
+        assert_eq!(
+            feature_type.attributes["array"].type_ref,
+            TypeRef::JsonString(Box::new(Attribute::new(TypeRef::String)))
+        );
+        assert_eq!(
+            feature_type.attributes["boolean"].type_ref,
+            TypeRef::Boolean
+        );
+        assert_eq!(feature_type.attributes["double"].type_ref, TypeRef::Double);
+        assert_eq!(
+            feature_type.attributes["integer"].type_ref,
+            TypeRef::Integer
+        );
+        assert_eq!(
+            feature_type.attributes["object"].type_ref,
+            TypeRef::JsonString(Box::new(Attribute::new(TypeRef::String)))
+        );
+        assert_eq!(feature_type.attributes["string"].type_ref, TypeRef::String);
+    }
+
+    #[test]
+    fn widens_numeric_and_null_types_deterministically() {
+        let first = properties(json!({
+            "all_null": null,
+            "nullable_boolean": null,
+            "number": 1
+        }));
+        let second = properties(json!({
+            "all_null": null,
+            "nullable_boolean": false,
+            "number": 1.25
+        }));
+        let mut builder = GeoJsonSchemaBuilder::default();
+
+        builder.observe_properties(Some(&first));
+        builder.observe_properties(Some(&second));
+        let feature_type = builder.finish();
+
+        assert_eq!(
+            feature_type.attributes["all_null"].type_ref,
+            TypeRef::String
+        );
+        assert_eq!(
+            feature_type.attributes["nullable_boolean"].type_ref,
+            TypeRef::Boolean
+        );
+        assert_eq!(feature_type.attributes["number"].type_ref, TypeRef::Double);
+    }
+
+    #[test]
+    fn widens_incompatible_types_to_string() {
+        let first = properties(json!({"mixed": true, "structured": [1]}));
+        let second = properties(json!({"mixed": 1, "structured": {"key": "value"}}));
+        let third = properties(json!({"structured": "plain"}));
+        let mut builder = GeoJsonSchemaBuilder::default();
+
+        builder.observe_properties(Some(&first));
+        builder.observe_properties(Some(&second));
+        builder.observe_properties(Some(&third));
+        let feature_type = builder.finish();
+
+        assert_eq!(feature_type.attributes["mixed"].type_ref, TypeRef::String);
+        assert_eq!(
+            feature_type.attributes["structured"].type_ref,
+            TypeRef::String
+        );
+    }
+}
